101% - Réctification soutenance
-> Création d'un dossier/fichier "Router.jsx" pour placer le router en dehors de l'index.js + importation de ce dernier dans l'index.js

-> Correction de l'erreur de rediréction du lien retour à l'accueil dans la page Error
