+++ src/index.js
@@ -1,33 +1,10 @@
 //Utils
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { createBrowserRouter, RouterProvider } from 'react-router-dom'
+import { RouterProvider } from 'react-router-dom'
 
-//Pages
-import Home from './pages/Home/Home'
-import Logement from './pages/Logement/Logement'
-import About from './pages/About/About'
-import Error from './pages/Error/Error'
-
-//Router
-const router = createBrowserRouter([
-  {
-    path: '/Kasa.fr',
-    element: <Home />,
-  },
-  {
-    path: '/logement/:id',
-    element: <Logement />,
-  },
-  {
-    path: '/about',
-    element: <About />,
-  },
-  {
-    path: '/*',
-    element: <Error />,
-  },
-])
+//Component
+import router from './components/Router/Router'
 
 //Render
 const root = ReactDOM.createRoot(document.getElementById('root'))
